+++ AppCore/CoreDependencies.py
@@ -1,14 +1,18 @@
+import atexit
+
 from AppCore import *
 from AppCore.Config import ConfigurationManager
 from AppCore.Data.CardSearchDataSource import *
+from AppCore.Data.RecentPublishedDataSource import *
 from AppCore.Image import *
 from AppCore.Network import *
+from AppCore.Observation.Events import ApplicationEvent
 from AppCore.Observation.ObservationTower import ObservationTower
 from AppCore.Resource import *
-from AppCore.Service import PlatformServiceProvider
+from AppCore.Service import PlatformServiceProvider, StringFormatter
 
 from .CoreDependencyProviding import CoreDependencyProviding
-from AppCore.Data.RecentPublishedDataSource import *
+
 
 class CoreDependencies(CoreDependencyProviding):
     def __init__(self):
@@ -18,9 +22,12 @@ class CoreDependencies(CoreDependencyProviding):
                                                                       self._observation_tower)
             self._image_resource_deployer = ImageResourceDeployer(self._configuration_manager,
                                                             self._observation_tower)
+            self._string_formatter = StringFormatter()
             self._image_resource_processor_provider = None
             self._api_client_provider = None
             self._image_source_provider = None
+
+            atexit.register(self.cleanup)
     
     @property
     def image_resource_deployer(self) -> ImageResourceDeployer:
@@ -37,15 +44,11 @@ class CoreDependencies(CoreDependencyProviding):
     @property
     def platform_service_provider(self) -> PlatformServiceProvider:
         return self._platform_service_provider
-
-    # @property
-    # def image_resource_processor_provider(self) -> ImageResourceProcessorProviding:
-    #     return self._image_resource_processor_provider
-
-    # @property
-    # def api_client_provider(self) -> APIClientProviding:
-    #     return self._api_client_provider
     
-    # @property
-    # def image_source_provider(self) -> CardImageSourceProviding:
-    #     return self._image_source_provider
+    @property
+    def string_formatter(self) -> StringFormatter:
+        return self._string_formatter
+    
+    def cleanup(self):
+        print("Performing cleanup before exit...")
+        self._observation_tower.notify(ApplicationEvent(ApplicationEvent.EventType.APP_WILL_TERMINATE))