+++ AppUI/UIComponents/Base/PublishHistoryTableViewController.py
@@ -1,4 +1,4 @@
-from PyQt5.QtWidgets import QListWidget, QVBoxLayout, QWidget
+from PyQt5.QtWidgets import QListWidget, QVBoxLayout, QWidget, QListWidgetItem
 
 from AppCore.Data.RecentPublishedDataSource import *
 from AppCore.Observation import *
@@ -17,6 +17,7 @@ class PublishHistoryTableViewController(QWidget, TransmissionReceiverProtocol, R
         self._image_preview_view = image_preview_view
         self._observation_tower = app_dependency_provider.observation_tower
         self._recent_published_data_source = recent_published_data_source
+        self._string_formatter = app_dependency_provider.string_formatter
         recent_published_data_source.delegate = self
         
         history_layout = QVBoxLayout()
@@ -28,11 +29,14 @@ class PublishHistoryTableViewController(QWidget, TransmissionReceiverProtocol, R
         
         self._observation_tower.subscribe_multi(self, [PublishStagedResourcesEvent])
     
+        self._update_history_list()
+
     def set_active(self):
         local_resource = self._recent_published_data_source.selected_local_resource
         if local_resource is not None:
             self._image_preview_view.set_image(local_resource)
-            self._observation_tower.notify(LocalResourceSelectedEvent(local_resource)) # rework?
+
+        # self._update_history_list()
     
     def rp_did_retrieve_card_resource_for_card_selection(self, ds: ..., local_resource: LocalCardResource) -> None:
         self.set_active()
@@ -45,7 +49,9 @@ class PublishHistoryTableViewController(QWidget, TransmissionReceiverProtocol, R
     def _update_history_list(self):
         self._history_list.clear()
         for r in self._recent_published_data_source.published_resources_history:
-            self._history_list.addItem(f'{r[1].strftime("%m/%d/%Y, %I:%M %p")} - {r[0].display_name}')
+            item = QListWidgetItem(f'{self._string_formatter.format_date(r[1])} - {r[0].display_name}')
+            item.setToolTip(f'{r[1].strftime("%m/%d/%Y, %I:%M %p")}')
+            self._history_list.addItem(item)
     
     def handle_observation_tower_event(self, event: TransmissionProtocol):
         if type(event) == PublishStagedResourcesEvent: