+++ AppUI/UIComponents/Base/__init__.py
@@ -2,3 +2,5 @@ from .ImagePreviewViewController import *
 from .SearchTableViewController import *
 from .AddImageCTAViewController import *
 from .PublishHistoryTableViewController import *
+from .ImagePreviewLocalResourceDataSourceDecorator import *
+from .SearchHistoryTableViewController import *